refactor: separating out endpoint and search compilation units
