+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/entities/finder/DatasetsQuery.scala
@@ -19,12 +19,12 @@
 package io.renku.knowledgegraph.entities
 package finder
 
+import Criteria.Filters.EntityType
 import cats.data.NonEmptyList
 import cats.syntax.all._
 import io.circe.{Decoder, DecodingFailure}
 import io.renku.graph.model._
 import io.renku.graph.model.images.ImageUri
-import io.renku.knowledgegraph.entities.Endpoint.Criteria.Filters.EntityType
 import io.renku.knowledgegraph.entities.model.{Entity, MatchingScore}
 
 private case object DatasetsQuery extends EntityQuery[model.Entity.Dataset] {
@@ -46,7 +46,7 @@ private case object DatasetsQuery extends EntityQuery[model.Entity.Dataset] {
 
   override val entityType: EntityType = EntityType.Dataset
 
-  override def query(criteria: Endpoint.Criteria) = (criteria.filters whenRequesting entityType) {
+  override def query(criteria: Criteria) = (criteria.filters whenRequesting entityType) {
     import criteria._
     // format: off
     s"""|{
