+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/entities/finder/WorkflowsEntitiesFinderSpec.scala
@@ -19,6 +19,7 @@
 package io.renku.knowledgegraph.entities
 package finder
 
+import Criteria.Filters
 import cats.syntax.all._
 import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators._
@@ -26,8 +27,6 @@ import io.renku.graph.model.GraphModelGenerators._
 import io.renku.graph.model._
 import io.renku.graph.model.testentities.RenkuProject.CreateCompositePlan
 import io.renku.graph.model.testentities._
-import io.renku.knowledgegraph.entities.Endpoint.Criteria
-import io.renku.knowledgegraph.entities.Endpoint.Criteria.Filters
 import io.renku.knowledgegraph.entities.model.Entity.Workflow.WorkflowType
 import io.renku.testtools.IOSpec
 import io.renku.triplesstore.{InMemoryJenaForSpec, ProjectsDataset}
@@ -177,10 +176,10 @@ class WorkflowsEntitiesFinderSpec
         .findEntities(Criteria(filters = Filters(entityTypes = Set(Filters.EntityType.Workflow))))
         .unsafeRunSync()
 
-      results.pagingInfo.total.value shouldBe (project.plans.size)
+      results.pagingInfo.total.value shouldBe project.plans.size
 
       val wfs = results.results.collect { case e: model.Entity.Workflow => e }
-      wfs should have size (project.plans.size)
+      wfs should have size project.plans.size
 
       wfs.map(_.workflowType) should contain theSameElementsAs List(WorkflowType.Step, WorkflowType.Composite)
     }
